refactor(tui): extract workspace buffer owner (W2)
workspace-editor.ts moves equivalently to workspace/buffer-io.ts; new
buffer-owner.ts uniquely owns the editor pool (LRU order, active path, dirty
and external-changed sets, per-buffer metadata, textarea instances, cursor,
find/goto/save-as bars) and the save/save-as/reload lifecycle. The controller
composes it through narrow ports (onStatus/onWritten/onReloaded/
onSaveAsTargetActivated/onOverwriteConfirm/onSaveAsOverwrite/onReloadConfirm/
onSaveStarted); dialog, ops bar, validation, and external-editor state remain
in the controller until their waves. sameFileStamp now lives in buffer-io as
the single stamp authority.
