+++ src/tui/workspace/buffer-io.ts
@@ -1,8 +1,8 @@
 import { lstat, mkdir, readFile, rename, unlink, writeFile } from "node:fs/promises";
 import type { Stats } from "node:fs";
 import { dirname } from "node:path";
-import { OVERSIZED_BYTES, PREVIEW_LINE_CAP, type WorkspaceFilePreview } from "./workspace/tree-data";
-import { assertProjectRelativePath } from "./workspace/paths";
+import { OVERSIZED_BYTES, PREVIEW_LINE_CAP, type WorkspaceFilePreview } from "./tree-data";
+import { assertProjectRelativePath } from "./paths";
 
 /**
  * Editor kernel for the Workspace page (Scope B / #62, milestone B3): the
@@ -37,6 +37,11 @@ export type EditableFileRead = {
 };
 
 export type FileStamp = { mtimeMs: number; ino: number };
+
+/** Whether two disk identities describe the same file version (mtime + inode). */
+export function sameFileStamp(left: FileStamp, right: FileStamp): boolean {
+  return left.mtimeMs === right.mtimeMs && left.ino === right.ino;
+}
 
 /**
  * Whether a preview is eligible for the editable textarea. Mirrors the B2
